Added Pong game implementation
This commit contains the file PongGame.java, which is the primary implementation of the Pong game. There are two paddles and a ball in the game. The paddles are controlled using the W/S and Up/Down arrow keys. The game maintains track of the score, and when a point is scored, the ball returns to the center of the screen.